+++ src/features/subject/components/SubjectManager.tsx
@@ -1,12 +1,13 @@
 "use client";
 
+import { PageTitle } from "../../../components/rpg/PageTitle";
 import { SubjectForm } from "./SubjectForm";
 import { SubjectList } from "./SubjectList";
 
 export function SubjectManager() {
   return (
     <section className="flex flex-col gap-6">
-      <h1 className="text-lg font-semibold">Subjects</h1>
+      <PageTitle>Subjects</PageTitle>
       <SubjectForm />
       <SubjectList />
     </section>